Restructuring frame after view from instructor
