+++ src/App.jsx
@@ -4,7 +4,9 @@ import { Route, Routes } from "react-router-dom"
 import { NavBar } from './components/navBar/NavBar.jsx'
 import { Home } from './components/homePage/Home.jsx'
 import { MyCloset } from './components/myCloset/MyCloset.jsx'
-import { Upload } from './components/upload/Upload.jsx'
+import { AddClothingType } from './components/addClothingType/AddClothingType.jsx'
+
+// import { Upload } from './components/upload/Upload.jsx'
 
 
 
@@ -15,7 +17,8 @@ export const App = () => {
       <NavBar />
       <Home />
       <MyCloset />
-      <Upload />
+      <AddClothingType />
+      {/* <Upload /> */}
     </>
   )
 }
