+++ src/components/addClothingType/addClothingType.css
@@ -1,0 +1,4 @@
+.titles{
+    display: flex;
+    justify-content: space-around;
+}
